[US_ND] Add missing cols to docstars_contacts config (Recidiviz/recidiviz-data#12868)
A bunch of the docstars_contacts files from April 2021 and before are failing to re-import in staging because they have more columns than are listed in the config for that file. This PR adds those cols and also adds an option to make the import_raw_files_to_sandbox script stricter.

GitOrigin-RevId: 5c32790728773ffb8da9b725a402ad43f12a1741