[Ingest] Allows for parallel raw data import / ingest view materialization (Recidiviz/recidiviz-data#12935)
Renames the serial-task-queue module to base-task-queue and adds an argument for setting the max_concurrent_dispatches arg.

Then sets that value to 5 for both raw data import and ingest view materialization queues. Ideally we'd be able to make these values even bigger, but let's start small and see how ingest handles it.

GitOrigin-RevId: 4f46f24e1d1220d831993ab6e5e97e6681ba464c