[BQ materialization] Remove usages of GcsfsDirectIngestFileType (part 4) (Recidiviz/recidiviz-data#13123)
Now that ingest only deals with one type of GCS file, raw data files, we can eliminate GcsfsDirectIngestFileType entirely. This is the final of 4 PRs that will fully eliminate references to that enum, baking in the assumption that any GCS file in the ingest bucket is a "raw data file".

This change includes removal of all remaining references to GcsfsDirectIngestFileType and deletes the enum itself.

Part of Recidiviz/recidiviz-data#11424.

GitOrigin-RevId: 93caff2a0f3dcd3856c81a5b4b24c9064b69e964