[BQ materialization] Eliminate IngestViewMaterializationArgsGeneratorDelegate (Recidiviz/recidiviz-data#13125)
Now that BQ materialization is fully shipped, there is only one implementation of IngestViewMaterializationArgsGeneratorDelegate and no longer a need for a delegate to the IngestViewMaterializationArgsGenerator class at all. This PR moves all the logic in the BQBasedMaterializationArgsGeneratorDelegate into the IngestViewMaterializationArgsGenerator itself.

This is the last cleanup PR to complete Recidiviz/recidiviz-data#11424!

Closes Recidiviz/recidiviz-data#11424.

GitOrigin-RevId: b20192ef188dd9b1d6a1ffcf06bd1a90dc75d0c6