Fixes date bug (Recidiviz/recidiviz-data#13181)
Recidiviz/recidiviz-data#13036 replaced violation_sessions with violation_responses, which doesn't have an earliest_available_date field. This field was still present in supervision_officer_office_metrics, causing the deploy to break. I replaced that field with response_date and filed Recidiviz/recidiviz-data#13180 to rename the violation metric fields to reflect that they are now counting responses and not violations.

GitOrigin-RevId: e92c67ad935228c65d3e68b557e4f00217ffcdec