[Testing] Remove unnecessarily set data_types in ingest view tests (Recidiviz/recidiviz-data#13579)
We have never holistically thought about how we handle type conversion when comparing fixture CSV files to data we have generated via an ingest view query against fake Postgres data.

This PR attempts to fully eliminate the need to manually specify data types when writing tests by instead inferring the types of the data in the query results DataFrame and either:
a) Converting data in the results dataframe to a type that is easier to compare against fixture data, or
b) Convert the types in the dataframe we get from the fixture files to better match the types in the results dataframe.

Unblocks Recidiviz/recidiviz-data#12568.

GitOrigin-RevId: ddc763f7ca1302e60d786239305516419528ed72