[Testing] Throw if raw data fixture has columns that do not match config (Recidiviz/recidiviz-data#13626)
In development yesterday, Genevieve discovered a fixture file for a TN ingest view test that had a column in it that was not listed in the raw file config (because that column does not exist in the real data).

This adds a check to enforce that raw data fixtures must match raw data configs.

This new check found one instance of bad columns in TN.

Follow-up to Recidiviz/recidiviz-data#13618.

GitOrigin-RevId: d307de5645adb3193183123371e68b001deac820