[Cleanup] Remove support for JAILS BQ refresh (Recidiviz/recidiviz-data#13725)
In Recidiviz/recidiviz-data#13705 we are going to delete the Cloud SQL instance for the JAILS schema, which will also delete the cloud SQL BQ connection used by the federated refresh. I don't think that deletion will cause a problem as long as we don't try to trigger a JAILS schema refresh, but I want to have this PR that rips JAILS out of the refresh ready just in case.

Part of Recidiviz/recidiviz-data#13703.

GitOrigin-RevId: 85f07eb9caacedb5f03a51b71ee5f1128732b057