[Admin Panel] Fix crash in Data Freshness page (Recidiviz/recidiviz-data#13752)
The data freshness page is currently crashing because there is a rerun currently in progress in PRIMARY for TN, which means that there are some ingest view tables with no rows that have processed dates before the last refresh. This is a valid case, so this change just updates the logic to allow for that.

Closes Recidiviz/recidiviz-data#13349

GitOrigin-RevId: ddf02e8f7263197e13fda53f9c5782bf331bfb4b