[Admin Panel] Allow developers without JC secrets access to develop locally (Recidiviz/recidiviz-data#13958)
Ola does not have access to load the Auth0 secrets in recidiviz/tools/justice_counts/control_panel/initialize_development_environment.sh. Ola doesn't actually need to do JC admin panel development, so this is fine, but the way we are instantiating the Auth0 client for JC means that it won't allow us to spin up the admin panel backend if we haven't loaded those secrets.

This updates the code to only instantiate an Auth0 client if we're actually going to use it (a pattern we use for client objects in our codebase other places as well).

Unblocks Recidiviz/recidiviz-data#13781

GitOrigin-RevId: 727b01fe5fea2aa7d70849b3fb4520c0a8a1ba28