[Orchestration] Run validations in-line in the DAG (Recidiviz/recidiviz-data#13973)
This PR re-introduces changes in the reverted Recidiviz/recidiviz-data#13608, making it so validations are run in-line in the Airflow DAG. This version updates the DAG logic to queue a cloud task that runs the validations and wait for that task to complete, rather than relying on the IAPHttpRequestOperator to queue the task, since we have not been able to figure out how to make that operator work for long-running HTTP requests.

This PR may be easiest to review by looking at the commits individually. The first commit is the commit that was reverted, the second is all the modifications I made on top of that commit.

Part of Recidiviz/recidiviz-data#13775

GitOrigin-RevId: b05070aac550dc8153815d16de645c3a8f7e6fb9