[Cleanup] Remove some references to JAILS ingest from recidiviz/persistence (Recidiviz/recidiviz-data#14259)
Deletes most of the places where the jails / aggregate schemas are referenced in the recidiviz/persistence directory in shared files. Will follow up with a PR that deletes most of the files that can now be removed from that package in full.

Does not yet delete the JAILS schema itself.

Part of Recidiviz/recidiviz-data#13703.

GitOrigin-RevId: af0f4369ceb0fd7c90934f39606054a905d323b6