[Cleanup] Remove county-specific functionality from Region (Recidiviz/recidiviz-data#14864)
More county cleanup. The Region object has a lot of scraper functionality baked in that we no longer use. The Region is really now a direct ingest only object and should be renamed / moved accordingly, but did not include in this PR so it was easier to review changes.

Part of Recidiviz/recidiviz-data#13703

GitOrigin-RevId: 604fe1c0f12b1ff949c5043fa0807bc04a69c823