[Cleanup] Remove unnecessary type SystemLevel (Recidiviz/recidiviz-data#14928)
The SystemLevel enum, which used to tell us whether we were doing ingest into the jails or state schema, now only has one value, STATE. This PR removed the enum entirely. In the future, if we were to want to ingest into different schemas using code in persistence.py, we could thread the SchemaType enum through to the appropriate places.

Related to Recidiviz/recidiviz-data#13703.

GitOrigin-RevId: 63ccbf59fcff4e040258b739ac7b0d450fc91674