[Validation] Add test that checks that all validation views are configured in a validation job (Recidiviz/recidiviz-data#15054)
Includes a pretty hacky but effective test that checks that all defined validation view builders in the validation_views dataset are actually registered in a validation job.

Had to exempt a number views which this test found and filed a task to either move or register those views.

Part of Recidiviz/recidiviz-data#15080.

GitOrigin-RevId: d17cc5b1204fecb0538e716f0a955d10b292c2a6