[BigQuery] Some fixes to make view query generation deterministic (Recidiviz/recidiviz-data#15246)
I am trying to make a version of load_views_to_sandbox that loads all views that have changed on your local branch (as compared to views deployed to staging). This doesn't work very well right now, because there are a number of views whose queries are not generated deterministically. This PR adds sorting in a handful of places where we aren't enforcing ordering of columns, filter clauses, etc.

Prework for Recidiviz/recidiviz-data#6722

GitOrigin-RevId: 624d9a153792b15eddcf9a22a87ded2e5787b311